+++ app/utils/__init__.py
@@ -1,3 +1,4 @@
 from .weight_exercise_helpers import verify_weight_exercise
 from .cardio_exercise_helpers import verify_cardio_exercise
 from .food_helpers import verify_food
+from .user_helpers import calculate_age, calculate_bmr_for_men, calculate_bmr_for_women, convert_height_to_cm
